fragmentize the note
